refactor(layout): split NotificationBell into hook + panel modules
NotificationBell.tsx was 326 lines tangling five unrelated concerns:
static icon/colour lookup tables, `timeAgo` formatting, unread-count
polling (with visibility-aware pause/resume), paginated list fetching
with optimistic mark-read/delete, and the dropdown JSX itself. Split
them so each file has a single job:

- `notifications/notificationMeta.ts` — NOTIFICATION_ICONS,
  NOTIFICATION_COLORS, and the timeAgo helper.
- `notifications/useNotifications.ts` — hook owning all runtime state,
  the visibility-aware unread-count poll, lazy first-page load on
  open, load-more pagination, and optimistic mark-read / mark-all /
  delete mutations.
- `notifications/NotificationItem.tsx` — pure row (icon + title +
  message + delete button), receives callbacks.
- `notifications/NotificationPanel.tsx` — pure dropdown panel that
  renders loading / error / empty / list / load-more states from
  props only. forwardRef so the parent can handle click-outside.
- `NotificationBell.tsx` — 91-line thin orchestrator: open/close
  toggle, click-outside, bell button with unread badge. Everything
  else is delegated.

No behaviour change. Unread polls stay at 30s, list stale threshold
stays at 30s, pagination semantics are preserved.

Verified: tsc, eslint, knip ({"issues":[]}), vitest (85 tests) all
pass locally.

Made-with: Cursor
